+++ src/lib/account_update.ts
@@ -1,18 +1,12 @@
 import {
   circuitArray,
-  circuitValue,
+  provable,
+  provablePure,
   cloneCircuitValue,
   memoizationContext,
   memoizeWitness,
 } from './circuit_value.js';
-import {
-  Field,
-  Bool,
-  Ledger,
-  Circuit,
-  Pickles,
-  AsFieldElements,
-} from '../snarky.js';
+import { Field, Bool, Ledger, Circuit, Pickles, Provable } from '../snarky.js';
 import { Types } from '../snarky/types.js';
 import { PrivateKey, PublicKey } from './signature.js';
 import { UInt64, UInt32, Int64, Sign } from './int.js';
@@ -579,7 +573,7 @@ type LazyProof = {
   args: any[];
   previousProofs: { publicInput: Field[]; proof: Pickles.Proof }[];
   ZkappClass: typeof SmartContract;
-  memoized: Field[][];
+  memoized: { fields: Field[]; aux: any[] }[];
   blindingValue: Field;
 };
 
@@ -1069,7 +1063,7 @@ class AccountUpdate implements Types.AccountUpdate {
   }
 
   static witness<T>(
-    type: AsFieldElements<T>,
+    type: Provable<T>,
     compute: () => { accountUpdate: AccountUpdate; result: T },
     { skipCheck = false } = {}
   ) {
@@ -1079,7 +1073,7 @@ class AccountUpdate implements Types.AccountUpdate {
       Types.AccountUpdate.sizeInFields()
     );
     type combinedType = { accountUpdate: Field[]; result: T };
-    let combinedType = circuitValue<combinedType>({
+    let combinedType = provable({
       accountUpdate: accountUpdateType,
       result: type,
     });
@@ -1127,7 +1121,7 @@ class AccountUpdate implements Types.AccountUpdate {
    * which also get witnessed
    */
   static witnessTree<T>(
-    resultType: AsFieldElements<T>,
+    resultType: Provable<T>,
     childLayout: AccountUpdatesLayout,
     compute: () => { accountUpdate: AccountUpdate; result: T },
     options?: { skipCheck: boolean }
@@ -1153,7 +1147,7 @@ class AccountUpdate implements Types.AccountUpdate {
     let n = childArray.length;
     for (let i = 0; i < n; i++) {
       accountUpdate.children.accountUpdates[i] = AccountUpdate.witnessTree(
-        circuitValue<null>(null),
+        provable(null),
         childArray[i],
         () => ({
           accountUpdate:
@@ -1402,7 +1396,7 @@ function addMissingSignatures(
   Thus, the transaction is fully constrained by the proof - the proof couldn't be used to attest to a different transaction.
  */
 type ZkappPublicInput = { accountUpdate: Field; calls: Field };
-let ZkappPublicInput = circuitValue<ZkappPublicInput>(
+let ZkappPublicInput = provablePure(
   { accountUpdate: Field, calls: Field },
   { customObjectKeys: ['accountUpdate', 'calls'] }
 );